Replace GithubContributorsList to be from vibe-storybook-components
