+++ package.json
@@ -121,7 +121,7 @@
     "react-window": "^1.8.7",
     "react-windowed-select": "^2.0.4",
     "style-inject": "^0.3.0",
-    "vibe-storybook-components": "^0.10.3"
+    "vibe-storybook-components": "^0.10.4-external-contributions-d08b8ba5a7.0"
   },
   "devDependencies": {
     "@babel/core": "^7.23.2",
