+++ src/storybook/stand-alone-documentaion/welcome/other-contributors-list/other-contributors-list.module.scss
@@ -1,0 +1,3 @@
+.contributorsList {
+  margin-top: var(--sb-spacing-between-section-items);
+}
