+++ src/storybook/stand-alone-documentaion/welcome/welcome.stories.mdx
@@ -5,7 +5,6 @@ import { Principles } from "./principles/principles";
 import { Contributors } from "./contributors/contributors";
 import { WelcomeHeader } from "./welcome-header/welcome-header";
 import { OtherContributorsList } from "./other-contributors-list/other-contributors-list";
-import { Link } from "vibe-storybook-components";
 
 <Meta title="Welcome" />
